Fix SpriteData LP64 pointer slots so SpriteSetColor hits +0x28
Named SpriteData fields after pBase/pVramData were 8 bytes wide on the
64-bit port, so SpriteSetColor (called every field sprite draw via
func_80075B08) wrote RGB into field_0x30/direction instead of the packed
color word at +0x28. Store PSX pointer slots as u32 like ActorData.

+++ src/slus_006.64/system/animation_scripts.c
@@ -120,7 +120,7 @@ void func_80021BCC(void* arg0, u32 arg1) {
 }
 
 void SpriteSetSpecialAnimFile(SpriteData* pSpriteData, void* pAnimFile) {
-    pSpriteData->pSpecialAnimFile = pAnimFile;
+    pSpriteData->pSpecialAnimFile = (u32)(uintptr_t)pAnimFile;
 }
 
 void func_80021BF8(void* arg0, s32 arg1) {
